pyproject.toml
[project]
name = "memories-tree"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.1"
dependencies = [
    "isort>=7.0.0",
    "langchain-core>=1.0.4",
    "langchain-openai>=1.0.2",
    "langgraph>=1.0.3",
    "pre-commit>=4.4.0",
    "pyright>=1.1.407",
    "pytest>=9.0.1",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.5",
]
authors = [
    {name = "Daniel Majewski",email = "danielmajewski@hotmail.com"},
]


[tool.ruff]
fix = true
line-length = 120
src = ["."]
extend-exclude = [".venv", "cdk.out", "**/dist", "**/__pycache__", ".ruff_cache"]
target-version = "py313"

[tool.ruff.lint]
ignore = ["E402", "ERA001"]
unfixable = [
    "ERA", # do not autoremove commented out code
]
extend-select = [
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "ERA", # flake8-eradicate/eradicate
    "I",   # isort
    "N",   # pep8-naming
    "PIE", # flake8-pie
    "PGH", # pygrep
    "RUF", # ruff checks
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    # "TID", # flake8-tidy-imports  # TODO reenable once we can remove relative imports
    "UP",  # pyupgrade
]
[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.isort]
force-single-line = true
lines-between-types = 1
lines-after-imports = 2

[tool.pyright]
exclude = ["**/__pycache__", ".ruff_cache", "cdk.out", "**/dist"]
reportMissingImports = true
reportMissingTypeStubs = false
pythonVersion = "3.13"
pythonPlatform = "Linux"
typeCheckingMode = "strict"
